Atualizados arquivos, criadas filas
Criadas as filas que faltavam. Sistema de configuracao atualizado para
funcionar conforme projetado. Corrigidos erros e warnings.

+++ LabFinal/enums.h
@@ -3,7 +3,7 @@
 
 typedef enum
 {
-  nenhuma = 0,
+  nenhuma = (ULONG)0,
   pes,
   painel,
   vidro
